[refactor] #238 mapper 메소드 data 패키지로 이동
