+++ app/src/main/java/com/runnect/runnect/domain/entity/MyDrawCourse.kt
@@ -1,0 +1,13 @@
+package com.runnect.runnect.domain.entity
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class MyDrawCourse(
+    val courseId: Int?,
+    val image: String?,
+    val city: String,
+    val region: String,
+    val title: String
+) : Parcelable
